Move scripts into language-specific folders, create Python package
