+++ scripts/js/prepublish.js
