+++ scripts/python/convert_band_data.py
