+++ scripts/python/get_cytobands_from_remote_dbs.py
@@ -9,15 +9,15 @@ Examples:
 # - Bonus: Convert this data into AGP 2.0, send data missing from NCBI to them
 
 import pymysql
-import urllib.request
 import os
 import json
 from concurrent.futures import ThreadPoolExecutor
 import logging
-import time
 import pprint
 import re
 import argparse
+
+from utils import request, connect, time_ms, natural_sort, chunkify
 
 parser = argparse.ArgumentParser(
     description=__doc__,
@@ -92,46 +92,6 @@ logger.addHandler(ch)
 time_ncbi = 0
 time_ucsc = 0
 time_ensembl = 0
-
-
-def request(url, request_body=None):
-    """Wrapper for urllib.request; includes caching
-    """
-    file_name = \
-        url.replace('.', '_').replace('/', '_').replace(':', '_')\
-        .replace('?', '_').replace('=', '_').replace('&', '_')
-
-    if fresh_run:
-        if request_body is not None:
-            req = urllib.request.Request(url, data=request_body)
-            data = urllib.request.urlopen(req).read().decode()
-        else:
-            with urllib.request.urlopen(url) as response:
-                data = response.read().decode('utf-8')
-        if fill_cache:
-            with open(cache_dir + file_name, 'w') as file:
-                file.write(data)
-    else:
-        with open(cache_dir + file_name) as file:
-            data = file.read().decode('utf-8')
-
-    return data
-
-
-def natural_sort(l):
-    """From https://stackoverflow.com/a/4836734
-    """
-    convert = lambda text: int(text) if text.isdigit() else text.lower()
-    alphanum_key = lambda key: [convert(c) for c in re.split('([0-9]+)', key)]
-    return sorted(l, key=alphanum_key)
-
-
-def time_ms():
-    return int(round(time.time() * 1000))
-
-
-def chunkify(lst, n):
-    return [lst[i::n] for i in range(n)]
 
 
 def update_bands_by_chr(bands_by_chr, chr, band_name, start, stop, stain):
